distance_to_origin da double a int

+++ Local_station.cpp
@@ -3,5 +3,5 @@
 
 using namespace std;
 
-Local_station::Local_station(double distance_origin, const std::string &name)
+Local_station::Local_station(int distance_origin, const std::string &name)
     : Station(distance_origin, name, "locale"), transit_tracks{2} {}
